week7 ver0 Remove unnecessary files
