--- .env
@@ -1,5 +1,0 @@
-HOST = "localhost"
-DB_USERNAME = "root"
-DB_PASSWORD = "love_taipei"
-DATABASE = "website_taipei"
-DB_PASSWORD_BACKUP = "12345678"
