+++ .gitignore
@@ -137,3 +137,6 @@ dmypy.json
 
 # Cython debug symbols
 cython_debug/
+
+# VSCode setting
+.vscode/
