+++ app.py
@@ -1,9 +1,9 @@
 from flask import *
 
-from api_folder.api_user import blueprint_user
-from api_folder.api_attractions import blueprint_attractions
-from api_folder.api_booking import blueprint_booking
-from api_folder.api_orders import blueprint_orders
+from controller.api_folder.api_user import blueprint_user
+from controller.api_folder.api_attractions import blueprint_attractions
+from controller.api_folder.api_booking import blueprint_booking
+from controller.api_folder.api_orders import blueprint_orders
 from module import get_connection
 
 app=Flask(__name__)
@@ -11,11 +11,7 @@ app.config["JSON_AS_ASCII"]=False
 app.config["TEMPLATES_AUTO_RELOAD"]=True
 app.static_folder="app_folder"
 
-# secret_key = random.randint(1000000000, 9999999999)
 app.json.ensure_ascii = False
-
-# from flask_cors import CORS
-# CORS(app)
 
 app.config['connection_pool'] = get_connection.access_db()
 app.register_blueprint(blueprint_user)
@@ -40,5 +36,4 @@ def thankyou():
 
 
 app.config["DEBUG"] = True
-# app.run(port=3000)
 app.run(host="0.0.0.0", port=3000)
