--- booking_backup.html
@@ -1,95 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}booking{% endblock %}
-
-{% block css %}
-<link rel="stylesheet" type="text/css" href="/static/css/booking.css"/>
-{% endblock %}
-
-{% block content %}
-<div class="frame">
-    <div class="titleName button bold" id="titleName"></div>
-    <div class="emptyMsg medium">目前沒有任何待預訂的行程</div>
-    <div class="innerFrame">
-        <div class="attractionFrame">
-            <!-- <div class="attractionInsertTarget">
-                <div class="attractionImg">
-                    <img class="attractionImgInsertTarget">
-                </div>
-                <div class="attractionInfo">
-                    <div class="attractionName bold cyanText">
-                        <div class="attractionNameLabel">台北一日遊：</div>
-                        <div class="attractionNameCnt">卡卡卡</div>
-                    </div>
-                    <div class="attractionDate">
-                        <div class="attractionDateLabel bold">日期：</div>
-                        <div class="attractionDateCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionTime">
-                        <div class="attractionTimeLabel bold">時間：</div>
-                        <div class="attractionTimeCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionPrice">
-                        <div class="attractionPriceLabel bold">費用：</div>
-                        <div class="attractionPriceCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionLoc">
-                        <div class="attractionLocLabel bold">地點：</div>
-                        <div class="attractionLocCnt medium">1231231</div>
-                    </div>
-                </div>
-                <div class="delImgFrame">
-                    <img class="delImg" src="/static/images/del.png">
-                </div>
-            </div> -->
-        </div>
-        <hr/>
-        <form class="userFrame">
-            <div class="userTitle button bold">您的聯絡資訊</div>
-            <div class="userName">
-                <label class="userNameLabel medium">聯絡姓名：</label>
-                <input class="userNameCnt medium inputbox" required type="text" name="userName">
-            </div>
-            <div class="userEmail">
-                <label class="userEmailLabel medium">連絡信箱：</label>
-                <input class="userEmailCnt medium inputbox" required type="email" name="userEmail">
-            </div>
-            <div class="userNumber">
-                <label class="userNumberLabel medium">手機號碼：</label>
-                <input class="userNumberCnt medium inputbox" maxlength="10" minlength="10" required type="text" inputmode="numeric" name="userNumber">
-            </div>
-            <div class="userReminder content bold">
-                請保持手機暢通，準時到達，導覽人員將用手機與您聯繫，務必留下正確的聯絡方式。
-            </div>
-        </form>
-        <hr/>
-        <form class="billFrame">
-            <div class="billTitle button bold">信用卡付款資訊</div>
-            <div class="billNumber">
-                <label class="billNumberLabel medium">卡片號碼：</label>
-                <input class="billNumberCnt medium inputbox" maxlength="16" minlength="16" required type="text" name="billNumber" placeholder="**** **** **** ****">
-            </div>
-            <div class="billExp">
-                <label class="billExpLabel medium">過期時間：</label>
-                <input class="billExpCnt medium inputbox" maxlength="5" required type="text" name="billExp" placeholder="MM / YY">
-            </div>
-            <div class="billPW">
-                <label class="billPWLabel medium">驗證密碼：</label>
-                <input class="billPWCnt medium inputbox" maxlength="3" minlength="3" required type="text" name="billPW" placeholder="CVV">
-            </div>
-        </form>
-        <hr/>
-        <div class="checkFrame">
-            <div class="checkPrice bold">
-                <div class="checkPriceLabel">總價：</div>
-                <div class="checkPriceCnt">1231231</div>
-            </div>
-            <div class="checkBtn button whiteText cyanBg">確認訂單並付款</div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
-{% block script %}
-<script src="/static/js/booking.js"></script>
-{% endblock %}
