+++ controller/api_folder/api_user.py
@@ -1,9 +1,8 @@
 from flask import ( Blueprint, request, jsonify)
 from datetime import datetime, timedelta
-# import random, string
 import jwt
 
-import module.flask_modules as flask_modules
+import module.operate_db as operate_db
 from controller import qry_para_set
 
 
@@ -11,7 +10,6 @@ blueprint_user = Blueprint('blueprint_user', __name__, url_prefix ="/api")
 
 
 secret_key = "oQAQByzarbxzNYSvuP3V4sdZyfKqxeHq"
-# secret_key = "".join(random.choice(string.ascii_letters + string.digits) for i in range(32))
 
 
 # 會員註冊
@@ -26,7 +24,7 @@ def signup():
             raise ValueError
         
         command_paras = qry_para_set.signup(input_data)
-        if not flask_modules.query_create(command_paras):
+        if not operate_db.query_create(command_paras):
             raise ValueError
         res = {
             "ok" : True
@@ -73,7 +71,7 @@ def signin():
             # JWT Module會自動驗證過期時間
             decoded_data = jwt.decode(token, secret_key, algorithms="HS256")
             command_paras = qry_para_set.verify(decoded_data)
-            data = flask_modules.query_fetch_one(command_paras)
+            data = operate_db.query_fetch_one(command_paras)
 
             if not data:
                 raise ValueError
@@ -98,7 +96,7 @@ def signin():
                 raise ValueError
             email = input_data.get("email")
             command_paras = qry_para_set.signin(email)
-            data = flask_modules.query_fetch_one(command_paras)
+            data = operate_db.query_fetch_one(command_paras)
             if not data or input_data.get("password") != data.get("password"):
                 raise ValueError
             payload = {
