--- data/db_init_program/sub_functions/module_data/.env
@@ -1,5 +1,0 @@
-HOST = "localhost"
-# username 在 module "os" 好像是保留字
-DB_USERNAME = "root"
-DB_PASSWORD = "love_taipei"
-DB_PASSWORD_BACKUP = "12345678"
