+++ module/combine_query.py
@@ -1,6 +1,3 @@
-from module import get_connection
-
-
 # decorator of CRUD
 def preprocessing(func):
     def wrapper(db_connection, db_cursor, paras):
@@ -73,48 +70,3 @@ def combine_delete_query(paras):
     if paras.get("where"):
         res += f" WHERE {paras['where']}"
     return res
-
-
-# C
-@get_connection.get_connection
-@preprocessing
-def query_create(db_connection, db_cursor, command, target) -> bool:
-    db_cursor.execute(command, target)
-    db_connection.commit()
-    return True
-
-
-# R_fetch_one
-@get_connection.get_connection
-@preprocessing
-def query_fetch_one(db_connection, db_cursor, command, target) -> dict:
-    db_cursor.execute("SET SESSION group_concat_max_len = 100000")
-    db_cursor.execute(command, target)
-    return db_cursor.fetchone()
-
-
-# R_fetch_all
-@get_connection.get_connection
-@preprocessing
-def query_fetch_all(db_connection, db_cursor, command, target) -> dict:
-    db_cursor.execute("SET SESSION group_concat_max_len = 100000")
-    db_cursor.execute(command, target)
-    return db_cursor.fetchall()
-
-
-# U
-@get_connection.get_connection
-@preprocessing
-def query_update(db_connection, db_cursor, command, target) -> bool:
-    db_cursor.execute(command, target)
-    db_connection.commit()
-    return True
-
-
-# D
-@get_connection.get_connection
-@preprocessing
-def query_del(db_connection, db_cursor, command, target) -> bool:
-    db_cursor.execute(command, target)
-    db_connection.commit()
-    return True
