+++ templates/thankyou.html
@@ -13,36 +13,6 @@
     <div class="emptyMsg medium">查無該訂單</div>
     <div class="innerFrame">
         <div class="attractionFrame">
-            <!-- <div class="attractionInsertTarget">
-                <div class="attractionImg">
-                    <img class="attractionImgInsertTarget">
-                </div>
-                <div class="attractionInfo">
-                    <div class="attractionName bold cyanText">
-                        <div class="attractionNameLabel">台北一日遊：</div>
-                        <div class="attractionNameCnt">卡卡卡</div>
-                    </div>
-                    <div class="attractionDate">
-                        <div class="attractionDateLabel bold">日期：</div>
-                        <div class="attractionDateCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionTime">
-                        <div class="attractionTimeLabel bold">時間：</div>
-                        <div class="attractionTimeCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionPrice">
-                        <div class="attractionPriceLabel bold">費用：</div>
-                        <div class="attractionPriceCnt medium">1231231</div>
-                    </div>
-                    <div class="attractionLoc">
-                        <div class="attractionLocLabel bold">地點：</div>
-                        <div class="attractionLocCnt medium">1231231</div>
-                    </div>
-                </div>
-                <div class="delImgFrame">
-                    <img class="delImg" src="/static/images/del.png">
-                </div>
-            </div> -->
         </div>
         <hr/>
         <div class="userFrame">
