--- test.json
@@ -1,12 +1,0 @@
-{
-  "prime": "7f52d31835282f4186fbeb327b1620b0b89879bf927e33fb5a3939f30c97e1ae",
-  "partner_key": "partner_PQvEeMp5qH0SG1Lj5kYgyaFROjL0wV5vj0kV3vNkl98jDCAFtufTwfuZ",
-  "merchant_id": "Ezpay99_EASY_WALLET",
-  "details": "TapPay Test",
-  "amount": 10000,
-  "cardholder": {
-    "phone_number": "+886912345678",
-    "name": "aaa",
-    "email": "test@gmail.com"
-  }
-}
